Clarify metric "Change Request Closure Ratio"
Previous definition of the metric was not clear, as it could be understand
that it is measuring something similar to BMI or REI metrics. Changes try
to shed more light on the specific added value of the metric and remove a
reference to the REI metric, which is calculated following a different
approach.

Signed-off-by: Luis Cañas-Díaz <lcanas@bitergia.com>